Fixed bug in subFeed, displayEmptyFeed
