+++ src/client/feed/SubFeed.js
@@ -170,7 +170,6 @@ class SubFeed extends React.Component {
         {isAuthHomeFeed && <LetsGetStarted />}
         {empty && <ScrollToTop />}
         <Feed
-          empty={empty}
           content={content}
           isFetching={isFetching}
           hasMore={hasMore}
